learning day endpoints added

+++ src/main/java/mif/vu/ikeea/Entity/Repository/LearningDayRepository.java
@@ -3,6 +3,6 @@ package mif.vu.ikeea.Entity.Repository;
 import mif.vu.ikeea.Entity.LearningDay;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LearningDayRepository extends CrudRepository<LearningDay, Integer> {
+public interface LearningDayRepository extends CrudRepository<LearningDay, Long> {
 
 }
